// src/runtime/Enums/supportedMimeTypes.ts
export enum SupportedMimeTypes {
    PDF = "application/pdf",
    JPEG = "image/jpeg",
    PNG = "image/png",
    GIF = "image/gif",
    SVG = "image/svg+xml",
    DOCX = "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
    XLS = "application/vnd.ms-excel",
    XLSX = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
    ACAD = "application/acad",
    DOC = "application/msword",
  }
  